# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Yastar API - multi-module business decision software
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /calculate/reverse-target:
    post:
      operationId: calculateReverseTarget
      tags: [calculate]
      summary: Reverse-calculate clients needed to hit a target profit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReverseTargetInput"
      responses:
        "200":
          description: Calculated result with feasibility check and insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReverseTargetResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /cost-items:
    get:
      operationId: listCostItems
      tags: [cost-items]
      summary: List the current owner's saved HPP cost items
      responses:
        "200":
          description: List of cost items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CostItem"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCostItem
      tags: [cost-items]
      summary: Save a new HPP cost item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CostItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostItem"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scenarios:
    get:
      operationId: listScenarios
      tags: [scenarios]
      summary: List the current owner's saved scenarios (all module types)
      responses:
        "200":
          description: List of scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scenario"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createScenario
      tags: [scenarios]
      summary: Save a new target_mundur scenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Scenario limit reached for current tier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scenarios/module:
    post:
      operationId: createModuleScenario
      tags: [scenarios]
      summary: Save a scenario for any module type (hpp, bep_usaha, harga_jual, etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleScenarioInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Scenario limit reached or module access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scenarios/{id}:
    get:
      operationId: getScenario
      tags: [scenarios]
      summary: Get a single saved scenario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateScenario
      tags: [scenarios]
      summary: Update/rename a saved scenario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioUpdate"
      responses:
        "200":
          description: Updated scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteScenario
      tags: [scenarios]
      summary: Delete a saved scenario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    BusinessType:
      type: string
      enum: [barbershop, salon, nail, spa, custom]
    CommissionModel:
      type: string
      enum: [flat, base_plus_commission, tiered]
...
    InsightSeverity:
      type: string
      enum: [info, success, warning, danger]
    Insight:
      type: object
      required: [severity, code, message]
      properties:
        severity:
          $ref: "#/components/schemas/InsightSeverity"
        code:
          type: string
        message:
          type: string
    ServiceItem:
      type: object
      required: [name, price, durationMinutes]
      properties:
        name:
          type: string
          minLength: 1
        price:
          type: number
          minimum: 0
        durationMinutes:
          type: integer
          minimum: 1
    CommissionTier:
      type: object
      required: [minClients, percent]
      properties:
        minClients:
          type: integer
          minimum: 0
        percent:
          type: number
          minimum: 0
          maximum: 100
    CommissionConfig:
      type: object
      properties:
        flatPercent:
          type: ["number", "null"]
          minimum: 0
          maximum: 100
        baseSalary:
          type: ["number", "null"]
          minimum: 0
        baseCommissionPercent:
          type: ["number", "null"]
          minimum: 0
          maximum: 100
        tiers:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/CommissionTier"
    ReverseTargetInput:
      type: object
      required:
        - businessType
        - employeeCount
        - workingDaysPerMonth
        - workingHoursPerDay
        - fixedCosts
        - targetProfit
        - commissionModel
        - commissionConfig
        - services
      properties:
        businessType:
          $ref: "#/components/schemas/BusinessType"
        employeeCount:
          type: integer
          minimum: 1
        workingDaysPerMonth:
          type: integer
          minimum: 1
          maximum: 31
        workingHoursPerDay:
          type: number
          minimum: 1
          maximum: 24
        fixedCosts:
          type: number
          minimum: 0
        targetProfit:
          type: number
          minimum: 0
        commissionModel:
          $ref: "#/components/schemas/CommissionModel"
        commissionConfig:
          $ref: "#/components/schemas/CommissionConfig"
        services:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ServiceItem"
    ReverseTargetResult:
      type: object
      required:
        - avgServicePrice
        - avgServiceDurationMinutes
        - effectiveCommissionPercent
        - netProfitPerClient
        - totalCostsMonthly
        - clientsNeededTotal
        - clientsNeededPerEmployee
        - clientsNeededPerDayPerEmployee
        - maxCapacityPerEmployeePerMonth
        - maxCapacityTotalPerMonth
        - utilizationPercent
        - marginPercent
        - isRealistic
        - insights
      properties:
        avgServicePrice:
          type: number
        avgServiceDurationMinutes:
          type: number
        effectiveCommissionPercent:
          type: number
        netProfitPerClient:
          type: number
        totalCostsMonthly:
          type: number
        clientsNeededTotal:
          type: number
        clientsNeededPerEmployee:
          type: number
        clientsNeededPerDayPerEmployee:
          type: number
        maxCapacityPerEmployeePerMonth:
          type: number
        maxCapacityTotalPerMonth:
          type: number
        utilizationPercent:
          type: number
        marginPercent:
          type: number
        isRealistic:
          type: boolean
        insights:
          type: array
          items:
            $ref: "#/components/schemas/Insight"
...
    CostItemInput:
      type: object
      required: [nama, mode, hppInput, hppResult]
      properties:
        nama:
          type: string
          minLength: 1
        mode:
          type: string
          enum: [produk, jasa]
        hppInput:
          type: object
          additionalProperties: true
        hppResult:
          type: object
          additionalProperties: true
    CostItem:
      type: object
      required: [id, nama, mode, hppInput, hppResult, createdAt, updatedAt]
      properties:
        id:
          type: integer
        nama:
          type: string
        mode:
          type: string
        hppInput:
          type: object
          additionalProperties: true
        hppResult:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ─── Scenarios ───────────────────────────────────────────────────────────
    Scenario:
      type: object
      required:
        - id
        - name
        - moduleType
        - businessType
        - employeeCount
        - workingDaysPerMonth
        - workingHoursPerDay
        - fixedCosts
        - targetProfit
        - commissionModel
        - commissionConfig
        - services
        - resultSnapshot
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        name:
          type: string
        moduleType:
          type: string
          default: target_mundur
        moduleInput:
          type: ["object", "null"]
          additionalProperties: true
        businessType:
          type: string
        employeeCount:
          type: integer
        workingDaysPerMonth:
          type: integer
        workingHoursPerDay:
          type: number
        fixedCosts:
          type: number
        targetProfit:
          type: number
        commissionModel:
          type: string
        commissionConfig:
          type: object
          additionalProperties: true
        services:
          type: array
          items:
            type: object
            additionalProperties: true
        resultSnapshot:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ScenarioInput:
      type: object
      required:
        - name
        - businessType
        - employeeCount
        - workingDaysPerMonth
        - workingHoursPerDay
        - fixedCosts
        - targetProfit
        - commissionModel
        - commissionConfig
        - services
      properties:
        name:
          type: string
          minLength: 1
        businessType:
          $ref: "#/components/schemas/BusinessType"
        employeeCount:
          type: integer
          minimum: 1
        workingDaysPerMonth:
          type: integer
          minimum: 1
          maximum: 31
        workingHoursPerDay:
          type: number
          minimum: 1
          maximum: 24
        fixedCosts:
          type: number
          minimum: 0
        targetProfit:
          type: number
          minimum: 0
        commissionModel:
          $ref: "#/components/schemas/CommissionModel"
        commissionConfig:
          $ref: "#/components/schemas/CommissionConfig"
        services:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ServiceItem"
    ScenarioUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        businessType:
          $ref: "#/components/schemas/BusinessType"
        employeeCount:
          type: integer
          minimum: 1
        workingDaysPerMonth:
          type: integer
          minimum: 1
          maximum: 31
        workingHoursPerDay:
          type: number
          minimum: 1
          maximum: 24
        fixedCosts:
          type: number
          minimum: 0
        targetProfit:
          type: number
          minimum: 0
        commissionModel:
          $ref: "#/components/schemas/CommissionModel"
        commissionConfig:
          $ref: "#/components/schemas/CommissionConfig"
        services:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ServiceItem"
    ModuleScenarioInput:
      type: object
      required: [name, moduleType, moduleInput, resultSnapshot]
      properties:
        name:
          type: string
          minLength: 1
        moduleType:
          type: string
          minLength: 1
        moduleInput:
          type: object
          additionalProperties: true
        resultSnapshot:
          type: object
          additionalProperties: true

    # ─── Account ─────────────────────────────────────────────────────────────
...