# lib/api-zod/src/generated/types/costItemInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Yastar API - multi-module business decision software
 * OpenAPI spec version: 0.2.0
 */
import type { CostItemInputHppInput } from './costItemInputHppInput';
import type { CostItemInputHppResult } from './costItemInputHppResult';
import type { CostItemInputMode } from './costItemInputMode';

export interface CostItemInput {
  /** @minLength 1 */
  nama: string;
  mode: CostItemInputMode;
  hppInput: CostItemInputHppInput;
  hppResult: CostItemInputHppResult;
}